Add files via upload
This was one of my assignments that I struggled with. The PDF was the assignment outline, the docx file is my planning stages (necessary for school and marks), and of course, the python file is what I did. The "Jennifer send back" py file is what my professor did to make the code work... I compared it both and realized I was kind of on the right track but not really. Lots of comments to help you understand.